refactor: Facade 패턴을 적용하여 리마인더 훅 계층 구조 최적화 (#2)
